+++ src/lib/components/v2/workflow/AddWorkflowTaskModal.svelte
@@ -19,6 +19,7 @@
 
 	let loading = false;
 	let addingTask = false;
+	let showDocLinksInTable = false;
 
 	/** @type {Array<import('fractal-components/types/api').TaskGroupV2>} */
 	let taskGroups = [];
@@ -94,6 +95,30 @@
 			}
 		);
 	}
+
+	/**
+	 * @param {import('fractal-components/types/api').TasksTableRow} taskRow
+	 */
+	function showTaskInfoButton(taskRow) {
+		return taskRow.docs_info || (!showDocLinksInTable && taskRow.docs_link);
+	}
+
+	/**
+	 * @param {import('fractal-components/types/api').TasksTableRow} taskRow
+	 */
+	function getTaskInfo(taskRow) {
+		let info = '';
+		if (taskRow.docs_info) {
+			info += taskRow.docs_info;
+		}
+		if (!showDocLinksInTable && taskRow.docs_link) {
+			if (taskRow.docs_info) {
+				info += '\n\n---\n'; // markdown hr
+			}
+			info += taskRow.docs_link;
+		}
+		return formatMarkdown(info);
+	}
 </script>
 
 <Modal
@@ -124,8 +149,8 @@
 				</svelte:fragment>
 				<svelte:fragment slot="extra-columns" let:task>
 					<td>
-						{#if task.docs_info}
-							<PropertyDescription description={formatMarkdown(task.docs_info)} html={true} />
+						{#if showTaskInfoButton(task)}
+							<PropertyDescription description={getTaskInfo(task)} html={true} />
 						{/if}
 					</td>
 					<td>