+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Added a "Get URL" button to image list (\#759);
 * Supported converter task types (\#759);
 * Made user sorting case insensitive (\#759);
+* Moved tasks table links to the info modal (\#759);
 
 # 1.16.2